move sensor data insertion to /scripts/data_loader
